requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Django==4.1.7
django-appconf==1.0.5
django-colorfield==0.8.0
django-cors-headers==3.14.0
django-editorjs==0.2.1
django-environ==0.10.0
django-templated-mail==1.1.1
django-unfold==0.15.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
idna==3.4
imagekitio==3.0.1
importlib-metadata==6.7.0
oauthlib==3.2.2
Pillow==9.5.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.15
whitenoise==6.4.0
zipp==3.16.2
